bugfix (NULL function pointer)

+++ src/test/cassoclist/1.c
@@ -71,7 +71,8 @@ int main(void)
 			fputs("Error!\n",stderr);
 			return -1;
 		}
-		fprintf(stdout,"add : %s,%d\n",integer_to_string(counter),pointer);
+		fprintf(stdout,"add : %s,%d\n"
+		    ,integer_to_string(counter),(unsigned int)pointer);
 		counter++;
 	}
 	counter = 0;
@@ -81,7 +82,8 @@ int main(void)
 		if(errcode){
 			fputs("Error!\n",stderr);
 		}
-		fprintf(stdout,"find : %s,%d\n",integer_to_string(counter),pointer);
+		fprintf(stdout,"find : %s,%d\n"
+		    ,integer_to_string(counter),(unsigned int)pointer);
 		counter++;
 	}
 	cassoclist_release(cassoclist);
